requirements.txt
matplotlib==3.7.1
numpy==1.23.5
#python==3.10.12
opencv-contrib-python==4.8.0.76
opencv-python-headless==4.8.1.78
torch==2.1.1
torchaudio==2.1.0
torchvision==0.16.0
mplstereonet
pandas
scikit-fuzzy
scikit-learn
Pillow
opencv-python
Flask
requests
boto3
